+++ internal/adaptor/handler/mock/file.go
@@ -3,7 +3,7 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=file.go -package=mock_transfer -destination=../../../adaptor/gateway/transfer/mock/file.go
+//	mockgen -source=file.go -package=mock_transfer -destination=../../../adaptor/handler/mock/file.go
 //
 // Package mock_transfer is a generated GoMock package.
 package mock_transfer
