+++ internal/adaptor/repository/mock/archive.go
@@ -3,7 +3,7 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=archive.go -package=mock_repository -destination=../../../adaptor/gateway/repository/mock/archive.go
+//	mockgen -source=archive.go -package=mock_repository -destination=../../../adaptor/repository/mock/archive.go
 //
 // Package mock_repository is a generated GoMock package.
 package mock_repository
