+++ internal/adaptor/repository/mock/client.go
@@ -3,7 +3,7 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=client.go -package=mock_repository -destination=../../../adaptor/gateway/repository/mock/client.go
+//	mockgen -source=client.go -package=mock_repository -destination=../../../adaptor/repository/mock/client.go
 //
 // Package mock_repository is a generated GoMock package.
 package mock_repository
