+++ internal/adaptor/repository/mock/device.go
@@ -3,7 +3,7 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=device.go -package=mock_repository -destination=../../../adaptor/gateway/repository/mock/device.go
+//	mockgen -source=device.go -package=mock_repository -destination=../../../adaptor/repository/mock/device.go
 //
 // Package mock_repository is a generated GoMock package.
 package mock_repository
