+++ internal/adaptor/repository/mock/evnet.go
@@ -3,7 +3,7 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=evnet.go -package=mock_repository -destination=../../../adaptor/gateway/repository/mock/evnet.go
+//	mockgen -source=evnet.go -package=mock_repository -destination=../../../adaptor/repository/mock/evnet.go
 //
 // Package mock_repository is a generated GoMock package.
 package mock_repository
