+++ internal/domain/gateway/repository/device.go
@@ -1,4 +1,4 @@
-//go:generate mockgen -source=$GOFILE -package=mock_repository -destination=../../../adaptor/gateway/repository/mock/$GOFILE
+//go:generate mockgen -source=$GOFILE -package=mock_repository -destination=../../../adaptor/repository/mock/$GOFILE
 package repository
 
 import (
